chore: rename binance loader to lower snake case
